move check request_page_size and response_next_page_token from AIP-158 to AIP-132
