+++ docs/rules/0158/response-plural-first-field.md
@@ -10,15 +10,13 @@ redirect_from:
 
 # Paginated methods: Page token field
 
-This rule enforces that all `List` and `Search` methods have a plural name
-repeatable field as a first field in the response message, as mandated in
-[AIP-158][].
+This rule enforces that Pagination requests have a plural name repeatable field
+as a first field in the response message, as mandated in [AIP-158][].
 
 ## Details
 
-This rule looks at any message matching `List*Response` or `Search*Response`
-that has `next_page_token` field and complains if the first field's name is not
-plural.
+This rule looks at any message that has `next_page_token` field and complains
+if the first field's name is not plural.
 
 ## Examples
